Se realizo resenasServie, falta probarlo Tambien cambie algunos atributos, en algunas clases, pot ejemplo si la clase pedido era un integer, en todas las clases habia que definirla como integer y no como long, lo mismo con long, una clase es long, su id siempre sera long.

+++ src/main/java/com/ecomarket/cl/ecomarket/repository/UsuarioRepository.java
@@ -7,11 +7,11 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
+public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
 
 
     Usuario findByCorreo(String correo);
 
 
-    Optional<Usuario> findById(Long id);  // Asegúrate de que este método esté presente
+    Optional<Usuario> findById(Integer id);  // Asegúrate de que este método esté presente
 }
